Use existing configuration
Passing in a path filename is not ideal when the configuration has already been loaded, when debugging a .net core website its looking for  "appsettings.json" which is not saved in the build directory. Also for ease of configuration profiles e.g. "appsettings.json", "appsettings.Development.json", "appsettings.Production.json" etc.